requirements.txt
certifi==2018.8.24
chardet==3.0.4
coverage==4.5.1
coveralls==1.5.1
Django==2.1.2
django-cors-middleware==1.2.0
django-extensions==1.7.1
django-filter==2.0.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
docopt==0.6.2
factory-boy==2.11.1
Faker==0.9.2
idna==2.7
Markdown==3.0.1
PyJWT==1.6.4
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
six==1.10.0
text-unidecode==1.2
urllib3==1.23
psycopg2
django-tinymce
geoip2
django-autoslug
django-admin-sortable